wip: refactor banner
