--- apps/wallet-mobile/src/TxHistory/TxHistoryList/RampOnOffBanner/strings.ts
@@ -1,20 +1,0 @@
-import {defineMessages} from 'react-intl'
-
-export const bannerRampOnOffMessages = defineMessages({
-  buyADA: {
-    id: 'rampOnOff.createRampOnOff.buyADA',
-    defaultMessage: '!!!Buy ADA',
-  },
-  getFirstAda: {
-    id: 'rampOnOff.rampOnOffScreen.getFirstAda',
-    defaultMessage: '!!!Welcome to Yoroi 👋️ Get your first ADA fast & easy',
-  },
-  ourTrustedPartners: {
-    id: 'rampOnOff.rampOnOffScreen.ourTrustedPartners',
-    defaultMessage: '!!!Top up your wallet safely using our trusted partners',
-  },
-  needMoreAda: {
-    id: 'rampOnOff.rampOnOffScreen.needMoreAda',
-    defaultMessage: '!!!Need more ADA for staking or swap?',
-  },
-})
