+++ apps/wallet-mobile/src/features/RampOnOff/common/MaybeShowBuyBanner/BuyBannerSmall.stories.tsx
@@ -1,0 +1,8 @@
+import {action} from '@storybook/addon-actions'
+import {storiesOf} from '@storybook/react-native'
+import React from 'react'
+
+import {ShowBuyBannerSmall} from './BuyBannerSmall'
+
+storiesOf('RampOnOff ShowBuyBannerSmall', module) //
+  .add('initial', () => <ShowBuyBannerSmall onClose={action('onClose')} />)
