+++ apps/wallet-mobile/src/features/RampOnOff/useCases/CreateExchange/CreateExchange.tsx
@@ -12,9 +12,9 @@ import {useTheme} from '../../../../theme'
 import {Theme} from '../../../../theme/types'
 import {useTokenInfo} from '../../../../yoroi-wallets/hooks'
 import {Quantities} from '../../../../yoroi-wallets/utils'
-import {useNavigateTo} from '../../common/navigation'
 import {useRampOnOff} from '../../common/RampOnOffProvider'
-import {useStrings} from '../../common/strings'
+import {useNavigateTo} from '../../common/useNavigateTo'
+import {useStrings} from '../../common/useStrings'
 import {EditAmount} from './EditAmount/EditAmount'
 import {SelectBuyOrSell} from './SelectBuyOrSell/SelectBuyOrSell'
 import {ShowDisclaimer} from './ShowDisclaimer/ShowDisclaimer'
