+++ apps/wallet-mobile/src/features/RampOnOff/useCases/CreateExchange/EditAmount/EditAmount.tsx
@@ -4,7 +4,7 @@ import {useSelectedWallet} from '../../../../../SelectedWallet'
 import {useBalance} from '../../../../../yoroi-wallets/hooks'
 import {AmountCard} from '../../../common/AmountCard/AmountCard'
 import {useRampOnOff} from '../../../common/RampOnOffProvider'
-import {useStrings} from '../../../common/strings'
+import {useStrings} from '../../../common/useStrings'
 
 export const EditAmount = () => {
   const wallet = useSelectedWallet()
