+++ apps/wallet-mobile/src/features/RampOnOff/useCases/CreateExchange/ShowProviderInfo/ShowProviderInfo.tsx
@@ -5,7 +5,7 @@ import banxaLogo from '../../../../../assets/img/banxa.png'
 import {Spacer} from '../../../../../components'
 import {useTheme} from '../../../../../theme'
 import {Theme} from '../../../../../theme/types'
-import {useStrings} from '../../../common/strings'
+import {useStrings} from '../../../common/useStrings'
 
 export const ShowProviderInfo = () => {
   const {theme} = useTheme()
