+++ apps/wallet-mobile/src/i18n/locales/en-US.json
@@ -880,9 +880,7 @@
   "rampOnOff.resultRampOnOff.and": "and",
   "rampOnOff.resultRampOnOff.customerSupport": "Yoroi Customer Support",
   "rampOnOff.resultRampOnOff.significant": "if you witnessed any significant transaction delays or errors.",
-  "rampOnOff.rampOnOffScreen.getFirstAda": "Welcome to Yoroi 👋️ Get your first ADA fast & easy",
-  "rampOnOff.rampOnOffScreen.needMoreAda": "Need more ADA for staking or swap?",
+  "rampOnOff.rampOnOffScreen.getFirstCrypto": "Welcome to Yoroi 👋️ Get your first ADA fast & easy",
+  "rampOnOff.rampOnOffScreen.needMoreCrypto": "Need more ADA for staking or swap?",
   "rampOnOff.rampOnOffScreen.ourTrustedPartners": "Top up your wallet safely using our trusted partners"
-
-
 }
